move coverage to test.js
